+++ CHANGES.md
@@ -9,6 +9,8 @@
 * Add ability to export data from a STAC catalog or API to files on disk.
 * Fix code that raised warnings from dependencies.
 * Log to stderr only by default and to a file only if requested.
+* Reorganize command line arguments to ensure shared options are always applied.
+* Remove option to call implementation scripts directly from the command line.
 
 ## [0.7.0](https://github.com/crim-ca/stac-populator/tree/0.7.0) (2025-03-07)